+++ lib/libmf/ffi.rb
@@ -53,7 +53,6 @@ module Libmf
     end
 
     attach_function :mf_get_default_param, [], Parameter.by_value
-    attach_function :mf_read_problem, [:string], Problem.by_value
     attach_function :mf_save_model, [Model.by_ref, :string], :int
     attach_function :mf_load_model, [:string], Model.auto_ptr
     attach_function :mf_destroy_model, [:pointer], :void