+++ test/test_helper.rb
@@ -20,8 +20,4 @@ class Minitest::Test
   def stress?
     ENV["STRESS"]
   end
-
-  def valgrind?
-    ENV["RUBY_MEMCHECK_RUNNING"]
-  end
 end