+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.5.0 (unreleased)
 
+- Dropped support for reading data directly from files
 - Dropped support for Ruby < 3.3
 
 ## 0.4.0 (2024-10-22)